requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pillow==10.2.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.40.0
datasets==2.13.1
scikit-learn==1.3.0
numpy==1.26.2
opencv-python==4.9.0.80
aiofiles==23.1.0
pydantic==1.10.11
python-dotenv==1.0.0
requests==2.31.0
matplotlib==3.8.1
pandas==2.2.2
